Fix width bug on mobile devices
The value reported by screen.availWidth is incorrect on my
Android device, causing the primary canvas element to occupy
less than a third of the screen. This effect is observed in
both Firefox and Chrome.
    To work around this, I have found that the browser window's
width is a reliable value to use. However, this value is not
a drop-in replacement, as on desktop systems the window's width
is variable. Taking the maximum between screen.availWidth and
the browser window width is a solution that works for both desktop
and mobile.