Correct mis-named variables
This is an odd case where the variables were incorrectly named
but were nevertheless used in the correct manner. I had mixed
up the names for the variables storing the number of rows and
columns for the canvas grid, yet strangely I was using them
correctly according to their value, as opposed to their name.